Triad fixes: producer-marked definite refusals, succeeded-based actor cleanliness, size-ratchet conformance (helpers to their domain modules)
