+++ ouroboros/agent.py
@@ -69,6 +69,8 @@ _worker_boot_lock = threading.Lock()
 from ouroboros.subagent_dispatch_notes import (  # noqa: E402
     dispatch_executor_note,
     executor_blocked_outcome,
+    _fill_executor_blocked_caps,
+    _nanny_route_dispatched_for,
 )
 
 
@@ -972,26 +974,16 @@ class OuroborosAgent:
         # The nanny postcondition's input fact for the loop's finalization seam:
         # THIS task was dispatched onto the delegated substrate. ALL economics
         # marks reset together per dispatch (F4) — defensive, since the
-        # ToolContext above is freshly built per task; see the helper.
-        # A configured agent_session row counts as dispatched even when the
-        # executor resolution reads "blocked": for a blocked start that is moot
-        # (the task terminals unrun below), but a mid-run failure must keep the
-        # reminders/nudges/chip alive on the wake loops (charter, 2026-08-28).
-        _snapshot = task.get("configured_subagent") if isinstance(
-            task.get("configured_subagent"), dict) else {}
-        _snap_route = _snapshot.get("route") if isinstance(_snapshot.get("route"), dict) else {}
-        reset_nanny_economics_marks(self.tools._ctx, route_dispatched=bool(
-            str(_snap_route.get("kind") or "") == "agent_session"
-            or (
-                dispatch is not None
-                and dispatch.executor_resolution is not None
-                and dispatch.executor_resolution.executor == "harness"
-            )
-        ), delegate_activity_seed=bool(
-            isinstance(getattr(ctx, "_configured_actor_bootstrap", None), dict)
-            and getattr(ctx, "_configured_actor_bootstrap", {}).get("physical_started")
-            or getattr(ctx, "_nanny_physical_activity_seed", False)
-        ))
+        # ToolContext above is freshly built per task; see the helpers.
+        reset_nanny_economics_marks(
+            self.tools._ctx,
+            route_dispatched=_nanny_route_dispatched_for(task, dispatch),
+            delegate_activity_seed=bool(
+                isinstance(getattr(ctx, "_configured_actor_bootstrap", None), dict)
+                and getattr(ctx, "_configured_actor_bootstrap", {}).get("physical_started")
+                or getattr(ctx, "_nanny_physical_activity_seed", False)
+            ),
+        )
 
         budget_remaining = None
         budget_accounting_status = "available"
@@ -1017,26 +1009,8 @@ class OuroborosAgent:
         # rather than a new return value or module-level helper, so synthesis can
         # adopt p34's `SubagentExecutorResolution`/`executor_blocked_outcome` without
         # a same-named twin to dedup here.
-        if dispatch is not None and dispatch.blocked and not startup_wake:
-            _res = dispatch.executor_resolution
-            cap_info["executor_blocked_reason"] = str(
-                (_res.reason if _res is not None else "")
-                or dispatch.delta.reason or "harness_not_configured"
-            )
-            cap_info["executor_blocked_requested"] = str(_res.requested if _res is not None else "harness")
-            cap_info["executor_blocked_reset_at"] = str(_res.reset_at if _res is not None else "")
-        elif not startup_wake and isinstance(
-            getattr(ctx, "_configured_startup_refusal", None), dict
-        ):
-            # Charter D2: the host's pre-start of a configured session leaf was
-            # DEFINITELY refused before the first model round (typed refusal, no
-            # custody handle) — the same $0 unrun terminal as a blocked pin.
-            _refusal = ctx._configured_startup_refusal
-            cap_info["executor_blocked_reason"] = str(
-                _refusal.get("reason") or "configured_session_unavailable"
-            )
-            cap_info["executor_blocked_requested"] = str(_refusal.get("requested") or "harness")
-            cap_info["executor_blocked_reset_at"] = str(_refusal.get("reset_at") or "")
+        if not startup_wake:
+            _fill_executor_blocked_caps(ctx, cap_info, dispatch)
         self._emit_live_log(
             "context_building_finished",
             task_id=str(task.get("id") or ""),
