+++ ouroboros/size_ratchet_manifest.py
@@ -207,7 +207,7 @@ BYTE_BASELINE_DEBT = {
 }
 
 BYTE_DEBT = {
-    "ouroboros/loop.py": 314553,
+    "ouroboros/loop.py": 314537,
     "tests/test_delegated_subagent_transport.py": 320571,
     "tests/test_devtools_benchmarks.py": 328282,
     "web/modules/chat.js": 225642,
